Slight changes in debug mode
